requirements.txt
boto3==1.37.34
botocore==1.37.34
numpy==2.2.4
pandas==2.2.3
pandas_market_calendars==4.4.0
polygon-api-client==1.14.5
pyarrow==19.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
s3transfer==0.11.4
setuptools==79.0.1
SQLAlchemy==2.0.40
mysql-connector-python==9.3.0


# Datafin package
-e ./datafin-package